restructure test configurations
