+++ packages/database/package.json
@@ -32,9 +32,9 @@
   "scripts": {
     "build": "tsc",
     "test": "yarn test:ci --detectOpenHandles",
-    "test:ci": "NODE_ENV=test jest --config ./tests/configs/ci.config.js --runInBand",
-    "test:k": "NODE_ENV=test jest --config ./tests/configs/khang.config.js --runInBand --detectOpenHandles",
-    "test:entity": "NODE_ENV=test jest --config ./tests/configs/entity.config.js --json --outputFile ./.logs/entity.json",
+    "test:ci": "NODE_ENV=test jest --config ./tests/configs/ci.config.ts --runInBand",
+    "test:k": "NODE_ENV=test jest --config ./tests/configs/khang.config.ts --runInBand --detectOpenHandles",
+    "test:pull": "NODE_ENV=test jest --config ./tests/configs/pull.config.ts --runInBand --detectOpenHandles",
     "lint": "eslint --fix --ext .js,.ts .",
     "lint:prettier": "prettier --ignore-path .gitignore --write . && yarn lint",
     "dump": "ts-node ./src/api/sql/dump.ts",
