--- packages/database/tests/configs/ci.config.js
@@ -1,6 +1,0 @@
-const base = require('./jest.config')
-
-module.exports = {
-  ...base,
-  silent: true,
-}
