+++ packages/database/tests/configs/ci.config.ts
@@ -1,0 +1,2 @@
+import { ci } from './jest.config'
+export default ci
