--- packages/database/tests/configs/entity.config.js
@@ -1,7 +1,0 @@
-const base = require('./jest.config')
-
-module.exports = {
-  ...base,
-  testMatch: ['**/tests/entity/*.test.ts'],
-  silent: false,
-}
