+++ packages/database/tests/configs/pull.config.ts
@@ -1,0 +1,2 @@
+import { pull } from './jest.config'
+export default pull
