+++ packages/database/tests/module-condensed.test.ts
@@ -1,7 +1,7 @@
-import { endpoint } from '../src/data-source'
+import { container, endpoint } from '../src/data-source'
 import { setup } from './setup'
-import { ModuleCondensed } from '../src/entity'
-import { remove } from '../src/sql'
+import { ModuleCondensed } from '../src/entity-repo/ModuleCondensed'
+import { ModuleCondensedRepository } from '../src/repository/ModuleCondensed'
 
 const lowerBound = 6000
 
@@ -11,13 +11,10 @@ beforeAll(async () => {
   await setup()
 })
 
-/* ===================================================================
- *   MODULE CONDENSED
- * ===================================================================
- */
-
 test('moduleCondensed.get', async () => {
-  const moduleList = await endpoint(ModuleCondensed.get)
+  const moduleList = await endpoint(
+    async () => await container(() => ModuleCondensedRepository.find())
+  )
   expect(moduleList).toBeDefined()
   if (!moduleList) {
     return
@@ -34,18 +31,25 @@ test('moduleCondensed.get', async () => {
 })
 
 test('moduleCondensed.getCodes', async () => {
-  const moduleList = await endpoint(ModuleCondensed.getCodes)
+  const moduleList = await endpoint(
+    async () => await container(() => ModuleCondensedRepository.getCodes())
+  )
   expect(moduleList).toBeDefined()
-  expect(moduleList).toBeInstanceOf(Set)
   if (!moduleList) {
     return
   }
-  expect(moduleList.size).toBeGreaterThan(lowerBound)
-  expect(moduleList.size).toStrictEqual(total)
+  expect(moduleList).toBeInstanceOf(Array)
+  moduleList.forEach((moduleCode) => {
+    expect(typeof moduleCode).toBe('string')
+  })
+  expect(moduleList.length).toBeGreaterThan(lowerBound)
+  expect(moduleList.length).toStrictEqual(total)
 })
 
 test('moduleCondensed.fetch', async () => {
-  const moduleList = await endpoint(ModuleCondensed.fetch)
+  const moduleList = await endpoint(
+    async () => await container(() => ModuleCondensedRepository.fetch())
+  )
   expect(moduleList).toBeDefined()
   if (!moduleList) {
     return
@@ -56,27 +60,6 @@ test('moduleCondensed.fetch', async () => {
   })
   const s = new Set(moduleList)
   expect(s.size).toBe(moduleList.length)
-  expect(s.size).toBeGreaterThan(lowerBound)
-  expect(s.size).toStrictEqual(total)
-})
-
-jest.setTimeout(10000)
-test('moduleCondensed.pull', async () => {
-  remove.tables(['moduleCondensed'])
-  const pullOnEmpty = await endpoint(ModuleCondensed.pull)
-  const pullOnFull = await endpoint(ModuleCondensed.pull)
-  const written = await endpoint(ModuleCondensed.get)
-
-  expect([pullOnFull, pullOnEmpty, written]).toBeDefined()
-  if (!pullOnFull || !pullOnEmpty || !written) {
-    return
-  }
-  /* make sure every element is a valid ModuleCondensed */
-  written.forEach((module) => {
-    expect(module).toBeInstanceOf(ModuleCondensed)
-  })
-  const s = new Set(written)
-  expect(s.size).toBe(written.length)
   expect(s.size).toBeGreaterThan(lowerBound)
   expect(s.size).toStrictEqual(total)
 })
