+++ packages/database/tests/module-pull.test.ts
@@ -1,48 +1,29 @@
-import { endpoint } from '../src/data-source'
-import { setup } from './setup'
-import { Module, ModuleCondensed } from '../src/entity'
+import { container, endpoint } from '../src/data-source'
+import { Module } from '../src/entity-repo/Module'
+import { ModuleRepository } from '../src/repository/Module'
 import { remove } from '../src/sql'
+import { setup } from './setup'
 
 beforeAll(async () => {
   await setup()
 })
 
-jest.setTimeout(120000)
-test('module.pull', async () => {
+async function pull() {
+  const res = await ModuleRepository.pull()
+  return res
+}
+
+jest.setTimeout(60000)
+test('pull all modules from NUSMods', async () => {
   await remove.tables(['degree_modules_required_module', 'module'])
-  // check that table removal was successfull
-  const mc = await endpoint(ModuleCondensed.getCodes)
-  const m = await endpoint(Module.getCodes)
-  expect(mc).toBeDefined()
-  expect(m).toBeDefined()
-  if (!mc || !m) {
+  const res = await endpoint(async () => await container(pull))
+  expect(res).toBeDefined()
+  expect(res).not.toBeNull()
+  if (!res) {
     return
   }
-  expect(m.size).toStrictEqual(0)
-
-  // wait for both to pull
-  const pm: void | Module[] = await endpoint(Module.pull)
-
-  // check pull validity
-  expect(pm).toBeDefined()
-  if (!pm) {
-    return
-  }
-  pm.forEach((module) => {
+  res.forEach((module) => {
     expect(module).toBeInstanceOf(Module)
   })
-  expect(pm.length).toBeGreaterThan(10)
-  expect(pm).toBeInstanceOf(Array)
-
-  if (!pm) {
-    return
-  }
-
-  /* make sure every element is a valid ModuleCondensed */
-  pm.forEach((module) => {
-    expect(module).toBeInstanceOf(Module)
-  })
-  const s = new Set(pm.map((x) => x.moduleCode))
-  expect(s.size).toBe(pm.length)
-  expect(s.size).toStrictEqual(mc.size)
+  expect(res.length).toBeGreaterThan(6100)
 })
